Rename window to tick, to prepare the rolling win.
The renaming is done so that the term "window" can be used as follow (to be
implemented later):

The sum will be computed over a window, which slides over ticks (frames?). A
new tick can enter the window, pushing an older tick out of the window. The sum
will be incremented by the entering tick and decremented by the outgoing tick.