Use JSON to represent cash register messages.
A single message from a cash register can contain multiple models/counts. This
is now done using JSON (maybe not the best serialization format but at least
this is uniform with the format used to communicate with the websocket client
(not already done)).

In addition, the ExclamationBolt class is removed.