Complete the `launch.sh` script.
It is not yet complete. Currently there are containers for one ZooKeeper
instance, one Kafka instance, a websocket server and a static web server
(Nginx). The image with the Storm topology must still be run manually. The next
step is to add containers to create a Storm cluster and run the topology on it.