Add additional bolt to compute N best models.
This bolt is similar to BestModelBolt but its output fields are of the same
type as RollingModelCountBolt, i.e. it can be fed to itself or to
BestModelBolt. This additional layer in the topology makes it possible to
overcome a possible bottleneck due to the single final BestModelBolt at the end
of the topology.